refactor(PatchNotes,PatchNotesModal): update imports to use json file
