+++ features/PatchNotes/components/PatchNotesModal.tsx
@@ -7,7 +7,7 @@ import {
   DialogTitle
 } from '@/shared/components/ui/dialog';
 import PostWrapper from '@/shared/components/layout/PostWrapper';
-import { patchNotesData } from '../patchNotesData';
+import patchNotesData from '../patchNotesData.json';
 
 interface PatchNotesModalProps {
   open: boolean;
@@ -20,11 +20,11 @@ export default function PatchNotesModal({
 }: PatchNotesModalProps) {
   return (
     <Dialog open={open} onOpenChange={onOpenChange}>
-      <DialogContent className='max-w-4xl w-[95vw] sm:w-[90vw] max-h-[85vh] sm:max-h-[80vh] flex flex-col p-0 gap-0'>
-        <DialogHeader className='px-6 pt-6 pb-4 border-b border-[var(--border-color)] sticky top-0 bg-[var(--background-color)] z-10'>
+      <DialogContent className='flex max-h-[85vh] w-[95vw] max-w-4xl flex-col gap-0 p-0 sm:max-h-[80vh] sm:w-[90vw]'>
+        <DialogHeader className='sticky top-0 z-10 border-b border-[var(--border-color)] bg-[var(--background-color)] px-6 pt-6 pb-4'>
           <DialogTitle className='text-2xl'>Patch Notes</DialogTitle>
         </DialogHeader>
-        <div className='overflow-y-auto px-6 py-4 flex-1'>
+        <div className='flex-1 overflow-y-auto px-6 py-4'>
           <div className='space-y-8'>
             {patchNotesData.map((patch, index) => (
               <div key={index}>
